+++ app/views/failedCommissionPayment.scala.html
@@ -1,0 +1,11 @@
+
+<div class="alert alert-block alert-warning fade in">
+	<h3>
+		<strong>Pago fallido</strong>
+	</h3>
+	<p>No se pudo realizar el pago correctamente.</p>
+	<p>Reintentá más tarde desde tu Dashboard.</p>
+	<p>
+		<a class="btn btn-warning" href="/#/dashboard">Volver al Dashboard</a> 
+	</p>
+</div>
