+++ app/views/successfulCommissionPayment.scala.html
@@ -1,0 +1,11 @@
+
+<div class="alert alert-block alert-info fade in">
+	<h3>
+		<strong>Pago exitoso!</strong>
+	</h3>
+	<p>
+		Tu pagos de comisiones están al día!
+	</p>
+	<a class="btn btn-info" href="/#/dashboard">Volver al Dashboard</a> 
+	</p>
+</div>
